+++ dcicutils/misc_utils.py
@@ -439,29 +439,83 @@ def as_seconds(*, seconds=0, minutes=0, hours=0, days=0, weeks=0, milliseconds=0
     return seconds
 
 
-HMS_TZ = pytz.timezone("US/Eastern")
+REF_TZ = pytz.timezone(os.environ.get("REF_TZ") or "US/Eastern")
 
 
-def as_datetime(dt, tz=None):
+class DatetimeCoercionFailure(ValueError):
+
+    def __init__(self, timespec, timezone):
+        self.timespec = timespec
+        self.timezone = timezone
+        extra = ""
+        if timezone:
+            extra = " (for timezone %s)" % timezone
+        super().__init__("Cannot coerce to datetime: %s%s" % (timespec, extra))
+
+
+def as_datetime(timespec, tz=None, raise_error=True):
     """
     Parses the given date/time (which may be a string or a datetime.datetime), returning a datetime.datetime object.
 
-    If the given datetime is already such an object, it is just returned.
+    If the given datetime is already such an object, it is just returned (not necessarily in the given timezone).
+    If the datetime to be returned has no timezone and a timezone argument has been given, that timezone is applied.
     If it is a string, it should be in a format such as 'yyyy-mm-dd hh:mm:ss' or 'yyyy-mm-dd hh:mm:ss-nnnn'
     (with -nnnn being a timezone specification).
+    If the given time is not a datetime, and cannot be coerced to be done, an error is raised
+    unless raise_error (default True) is False.
     """
-    if dt is None:
-        return None
     try:
         # This type check has to work even if datetime is mocked, so we use it under another variable name to
         # make it harder to mock out. -kmp 6-Nov-2020
+        dt = timespec
         if not isinstance(dt, datetime_type):
             dt = dateutil_parse(dt)
         if tz and not dt.tzinfo:
             dt = tz.localize(dt)
         return dt
     except Exception:
-        return None
+        # I decided to treat the returning None case as a bug. It was not advertised and not used.
+        # Throwing an error by default will make this more consistent with as_ref_datetime and as_utc_datetime.
+        # But just in case there is a use that wanted None, so it's easy to fix, raise_error=False can be supplied.
+        # -kmp 29-Nov-2020
+        if raise_error:
+            raise DatetimeCoercionFailure(timespec=timespec, timezone=tz)
+        else:
+            return None
+
+
+def as_ref_datetime(timespec):
+    """
+    Parses a given datetime, returning a rendition of that tie in the reference timezone (US/Eastern by default).
+
+    If the input time is a string or a naive datetime with no timezone, it is assumed to be in the reference timezone
+    (which is US/Eastern by default).
+    If the time is already a datetime, no parsing occurs, but the time is still adjusted to use the reference timeszone.
+    If the given time is not a datetime, and cannot be coerced to be done, an error is raised.
+    """
+    try:
+        dt = as_datetime(timespec, tz=REF_TZ)
+        hms_dt = dt.astimezone(REF_TZ)
+        return hms_dt
+    except Exception:
+        raise DatetimeCoercionFailure(timespec=timespec, timezone=REF_TZ)
+
+
+def as_utc_datetime(timespec):
+    """
+    Parses a given datetime, returning a rendition of that tie in UTC.
+
+    If the input time is a string or a naive datetime with no timezone, it is assumed to be in the reference timezone
+    (which is US/Eastern by default). UTC is only used as the output format, not as an assumption about the input.
+    If the time is already a datetime, no parsing occurs, but the time is still adjusted to use UTC.
+    If the given time is not a datetime, and cannot be coerced to be done, an error is raised.
+    """
+    try:
+        dt = as_datetime(timespec, tz=REF_TZ)
+        utc_dt = dt.astimezone(pytz.UTC)
+        return utc_dt
+    except Exception:
+        raise DatetimeCoercionFailure(timespec=timespec, timezone=pytz.UTC)
 
 
 def in_datetime_interval(when, *, start=None, end=None):
@@ -469,15 +523,20 @@ def in_datetime_interval(when, *, start=None, end=None):
     Returns true if the first argument ('when') is in the range given by the other arguments.
 
     The comparison is upper- and lower-inclusive.
+    The string will be parsed as a datetime in the reference timezone (REF_TZ) if it doesn't have an explicit timezone.
     """
-    start = as_datetime(start)
-    end = as_datetime(end)
+    when = as_ref_datetime(when)  # This is not allowed to be None, but it might be a str, and we need datetimes to compare.
+    start = start and as_ref_datetime(start)
+    end = end and as_ref_datetime(end)
     return (not start or start <= when) and (not end or end >= when)
 
 
-def hms_now():
-    """Returns the current time in the HMS timezone (US/Eastern, which may be EST or EDT, depending on time of year)."""
-    return as_datetime(datetime.datetime.now(), HMS_TZ)
+def ref_now():
+    """Returns the current time in the portal's reference timezone, as determined by REF_TZ.
+
+       Because this software originates at Harvard Medical School, the reference timezone defaults to US/Eastern.
+       It can be set to another value by binding the REF_TZ environment variable."""
+    return as_datetime(datetime.datetime.now(), REF_TZ)
 
 
 class LockoutManager:
@@ -999,3 +1058,8 @@ def getattr_customized(thing, key):
         raise UncustomizedInstance(instance=thing, field=key)
     else:
         return value
+
+
+# Deprecated names, still supported for a while.
+HMS_TZ = REF_TZ
+hms_now = ref_now